refactor(db): extract 8 modules incl. agents+channels — decomposition Wave 3
Wave 3: agents(+templates), channels, agent-profiles, node-profiles, savepoints,
plugins, health, webhooks. Bodies verbatim; barrel re-exports via `export *`.
channels (deferred in Wave 2) now wires acyclically — agents is extracted this wave.

- webhooks: moved master's REAL dispatchWebhook (guardedFetch redirect-hop SSRF
  re-check + deliverWithRetry), not the blueprint's older assertPublicHost version;
  removed the now-dead crypto/ssrf-guard imports from db.js's header (webhooks was
  their sole consumer).
- export* scope trap: 16 resident bare-calls -> 5 explicit import lines into db.js
  (agents/channels/plugins/savepoints/node-profiles). Acyclic.

db.js: 3159 -> 2042 (-1117). 23 of 31 modules now extracted. Gate: manifest --check
green (308); suite 45 files / 310.

Executor: GLM-5.2 (agentic, gate-driven). Reviewed + verified by m5Max.
