+++ server/db.js
@@ -1,6 +1,4 @@
 // =============== MYCELIUM — Database Layer ===============
-import crypto from 'crypto';
-import { assertPublicHost, guardedFetch, SSRFBlockedError } from './lib/ssrf-guard.js';
 import { db, stmt, buildUpdate, initDBConnection, getDB, DB_PATH } from './db/core.js';
 // Resident db.js functions (getBootPayload, overview/health) still call these now-extracted
 // helpers by bare name. The `export *` barrel below re-exports them to consumers but does NOT
@@ -15,6 +13,11 @@ import { getProjectConcepts, listConcepts, getConceptProjects } from './db/conce
 import { listPendingApprovalsByAgent, listApprovals } from './db/approvals.js';
 import { listPendingRequests, getUnreadMessages, markMessagesRead, listMessages, listTeamChat } from './db/messages.js';
 import { getContext, getAllContext, listContextKeys, getContextKey, upsertContextKey } from './db/context.js';
+import { getAgent, listAgents, resolveAssignee, updateAgentHeartbeat } from './db/agents.js';
+import { ensureDefaultChannels, getChannelsByUser, getUnreadCounts, isChannelMember, listChannels } from './db/channels.js';
+import { listPluginRecords } from './db/plugins.js';
+import { getLatestSavepoint } from './db/savepoints.js';
+import { buildCalibrationBlock, createNodeProfile, getNodeProfile, seedPlatformProfiles, updateNodeProfile } from './db/node-profiles.js';
 
 export { getDB };
 
@@ -33,6 +36,14 @@ export * from './db/approvals.js';
 export * from './db/runs.js';
 export * from './db/messages.js';
 export * from './db/context.js';
+export * from './db/agents.js';
+export * from './db/savepoints.js';
+export * from './db/plugins.js';
+export * from './db/health.js';
+export * from './db/webhooks.js';
+export * from './db/channels.js';
+export * from './db/agent-profiles.js';
+export * from './db/node-profiles.js';
 
 export function initDB() {
   initDBConnection();
@@ -43,72 +54,6 @@ export function initDB() {
 }
 
 // =============== MYCELIUM PLATFORM ===============
-
-// -- Agents --
-
-export function createAgent(id, name, projectId, apiKeyHash, capabilities) {
-  stmt('dvCreateAgent', `INSERT INTO agents (id, name, project_id, api_key_hash, capabilities)
-    VALUES (?, ?, ?, ?, ?)`).run(id, name, projectId, apiKeyHash, capabilities || '[]');
-}
-
-export function getAgent(id) {
-  return stmt('dvGetAgent', 'SELECT * FROM agents WHERE id = ?').get(id);
-}
-
-export function getAgentByKeyHash(apiKeyHash) {
-  return stmt('dvGetAgentByKey', 'SELECT * FROM agents WHERE api_key_hash = ?').get(apiKeyHash);
-}
-
-// Resolve an assignee to a canonical agent id. Accepts an agent id OR display
-// name, case-insensitively — so the app, a human, or a script can assign by name
-// ("Ada") and it still matches the agent whose id is "ada". Unmatched assignees
-// (operator names, not-yet-registered agents, "", null) pass through unchanged.
-export function resolveAssignee(assignee) {
-  if (typeof assignee !== 'string') return assignee;
-  var a = assignee.trim();
-  if (!a) return assignee;
-  var row = stmt('dvResolveAssignee',
-    'SELECT id FROM agents WHERE id = ? COLLATE NOCASE OR name = ? COLLATE NOCASE LIMIT 1'
-  ).get(a, a);
-  return row ? row.id : assignee;
-}
-
-export function listAgents() {
-  return stmt('dvListAgents4', "SELECT id, name, project_id, status, working_on, last_heartbeat, capabilities, avatar_url, role, operator_id, project, llm_backend, llm_model, runtime, system_diagnostics, agent_type, primary_team_id, created_at FROM agents WHERE project_id != 'drone' ORDER BY created_at").all();
-}
-
-export function listAllAgentsIncludingDrones() {
-  return stmt('dvListAllAgents', 'SELECT id, name, project_id, status, working_on, last_heartbeat, capabilities, avatar_url, role, operator_id, project, created_at FROM agents ORDER BY created_at').all();
-}
-
-export function updateAgentHeartbeat(id, status, workingOn) {
-  stmt('dvHeartbeat', `UPDATE agents SET status = ?, working_on = ?, last_heartbeat = datetime('now')
-    WHERE id = ?`).run(status || 'online', workingOn || '', id);
-}
-
-export function updateAgentKey(id, apiKeyHash) {
-  stmt('dvUpdateAgentKey', 'UPDATE agents SET api_key_hash = ? WHERE id = ?').run(apiKeyHash, id);
-}
-
-export function deleteAgent(id) {
-  db.prepare('DELETE FROM tasks WHERE assignee = ?').run(id);
-  db.prepare('DELETE FROM messages WHERE from_agent = ? OR to_agent = ?').run(id, id);
-  db.prepare('DELETE FROM bugs WHERE assignee = ?').run(id);
-  db.prepare('DELETE FROM drone_jobs WHERE requester = ? OR drone_id = ?').run(id, id);
-  db.prepare('DELETE FROM agent_savepoints WHERE agent_id = ?').run(id);
-  db.prepare('DELETE FROM webhooks WHERE agent_id = ?').run(id);
-  db.prepare('DELETE FROM message_reads WHERE agent_id = ?').run(id);
-  db.prepare('DELETE FROM channel_members WHERE user_id = ? AND user_type = ?').run(id, 'agent');
-  db.prepare('DELETE FROM agents WHERE id = ?').run(id);
-}
-
-export function updateAgent(id, fields) {
-  // Pre-process JSON fields
-  if (fields.system_diagnostics !== undefined && typeof fields.system_diagnostics !== 'string') {
-    fields = Object.assign({}, fields, { system_diagnostics: JSON.stringify(fields.system_diagnostics) });
-  }
-  buildUpdate('agents', id, fields, ['avatar_url', 'name', 'role', 'operator_id', 'project', 'project_id', 'llm_backend', 'llm_model', 'agent_type', 'capabilities', 'system_diagnostics', 'runtime']);
-}
 
 // -- Tasks --
 
@@ -296,23 +241,6 @@ export function deleteAsset(id) {
 
 export function listAssetsByDroneJob(droneJobId) {
   return db.prepare('SELECT * FROM assets WHERE drone_job_id = ?').all(droneJobId);
-}
-
-export function cleanupSavepoints(keepPerAgent) {
-  var keep = keepPerAgent || 50;
-  var agents = db.prepare("SELECT DISTINCT agent_id FROM agent_savepoints").all();
-  var totalCleaned = 0;
-  for (var i = 0; i < agents.length; i++) {
-    var agentId = agents[i].agent_id;
-    var result = db.prepare(
-      "DELETE FROM agent_savepoints WHERE agent_id = ? AND id NOT IN (SELECT id FROM agent_savepoints WHERE agent_id = ? ORDER BY heartbeat_at DESC LIMIT ?)"
-    ).run(agentId, agentId, keep);
-    totalCleaned += result.changes;
-  }
-  if (totalCleaned > 0) {
-    console.log('[mycelium] Cleaned up %d old savepoints (keep: %d per agent)', totalCleaned, keep);
-  }
-  return totalCleaned;
 }
 
 // -- Widgets --
@@ -1209,340 +1137,6 @@ export function completeLinkedPlanSteps(taskId) {
 
 // -- Studio Users --
 
-// -- Webhooks --
-
-export function createWebhook(agentId, url, events, secret) {
-  var eventsJson = Array.isArray(events) ? JSON.stringify(events) : (events || '["task_created","request_created","message_sent"]');
-  var result = db.prepare(
-    "INSERT INTO webhooks (agent_id, url, events, secret) VALUES (?, ?, ?, ?) RETURNING id"
-  ).get(agentId, url, eventsJson, secret || '');
-  return result.id;
-}
-
-export function listWebhooks(agentId) {
-  if (agentId) {
-    return db.prepare("SELECT * FROM webhooks WHERE agent_id = ? AND active = 1").all(agentId);
-  }
-  return db.prepare("SELECT * FROM webhooks WHERE active = 1").all();
-}
-
-export function deleteWebhook(id) {
-  db.prepare("DELETE FROM webhooks WHERE id = ?").run(id);
-}
-
-export async function dispatchWebhook(event, agentId, data) {
-  // Query webhooks for the target agent AND __global__ (admin-claude receives all events)
-  var webhooks = db.prepare(
-    "SELECT * FROM webhooks WHERE active = 1 AND (agent_id = ? OR agent_id = '__global__')"
-  ).all(agentId);
-
-  for (var wh of webhooks) {
-    var events = [];
-    try { events = JSON.parse(wh.events); } catch (e) { console.warn('[mycelium] JSON parse failed for webhook.events (webhook: ' + wh.id + '):', e.message); continue; }
-    if (events.indexOf(event) === -1 && events.indexOf('*') === -1) continue;
-
-    // Allow loopback if MYCELIUM_WEBHOOK_ALLOW_LOOPBACK is set to '1'. Declared at
-    // loop-body scope so the redirect-guarded delivery below reuses the same flag
-    // (the guardedFetch call sits outside the try, so a try-scoped const is undefined there).
-    const allowLoopback = process.env.MYCELIUM_WEBHOOK_ALLOW_LOOPBACK === '1';
-
-    // SSRF Guard: validate the webhook URL before making the request
-    try {
-      await assertPublicHost(wh.url, { allowLoopback });
-    } catch (ssrfError) {
-      if (ssrfError instanceof SSRFBlockedError) {
-        // Log the SSRF attempt and skip delivery
-        console.warn('[webhook] SSRF attempt blocked for webhook ' + wh.id + ':', ssrfError.message);
-        // Record the event for monitoring
-        try {
-          const payload = JSON.stringify({
-            event: event,
-            agent_id: agentId,
-            data: data,
-            timestamp: new Date().toISOString()
-          });
-          logWebhookDelivery(wh.id, event, agentId, payload, null, null, 'SSRF blocked: ' + ssrfError.message, 0);
-        } catch (logError) {
-          console.error('[webhook] Failed to log SSRF attempt:', logError.message);
-        }
-        continue; // Skip this webhook delivery
-      } else {
-        // Re-throw if it's not an SSRF error
-        throw ssrfError;
-      }
-    }
-
-    var payload = JSON.stringify({
-      event: event,
-      agent_id: agentId,
-      data: data,
-      timestamp: new Date().toISOString()
-    });
-
-    var headers = { 'Content-Type': 'application/json' };
-    if (wh.secret) {
-      var sig = crypto.createHmac('sha256', wh.secret).update(payload).digest('hex');
-      headers['X-Webhook-Signature'] = sig;
-    }
-
-    var whId = wh.id;
-    var startTime = Date.now();
-
-    // Non-blocking fetch with 5s timeout and retry (up to 3 attempts).
-    // guardedFetch re-runs the SSRF guard on every redirect hop so a public
-    // first-hop can't 302 into an internal address.
-    (function deliverWithRetry(url, opts, attempt) {
-      guardedFetch(url, Object.assign({}, opts, { signal: AbortSignal.timeout(5000) }), { allowLoopback })
-        .then(function (resp) {
-          var duration = Date.now() - startTime;
-          return resp.text().then(function (body) {
-            logWebhookDelivery(whId, event, agentId, payload, resp.status, body.substring(0, 1000), null, duration);
-          });
-        }).catch(function (err) {
-          var duration = Date.now() - startTime;
-          // SSRF blocks (incl. a redirect to a private host) must NOT be retried.
-          if (err instanceof SSRFBlockedError) {
-            logWebhookDelivery(whId, event, agentId, payload, null, null, 'SSRF blocked: ' + err.message, duration);
-            console.warn('[webhook] SSRF blocked for webhook ' + whId + ':', err.message);
-            return;
-          }
-          if (attempt < 3) {
-            var delay = Math.pow(2, attempt) * 1000; // 2s, 4s backoff
-            setTimeout(function () { deliverWithRetry(url, opts, attempt + 1); }, delay);
-          } else {
-            logWebhookDelivery(whId, event, agentId, payload, null, null, err.message + ' (after 3 attempts)', duration);
-            console.error('[webhook] Failed after 3 attempts to', url, ':', err.message);
-          }
-        });
-    })(wh.url, { method: 'POST', headers: headers, body: payload }, 1);
-  }
-}
-
-function logWebhookDelivery(webhookId, event, agentId, payload, statusCode, responseBody, error, durationMs) {
-  try {
-    db.prepare(
-      "INSERT INTO webhook_deliveries (webhook_id, event, agent_id, payload, status_code, response_body, error, duration_ms) VALUES (?, ?, ?, ?, ?, ?, ?, ?)"
-    ).run(webhookId, event, agentId, payload, statusCode, responseBody, error, durationMs);
-  } catch (e) {
-    console.error('[webhook-log] Failed to log delivery:', e.message);
-  }
-}
-
-export function listWebhookDeliveries(filters) {
-  var where = ['1=1'];
-  var params = [];
-  if (filters.event) { where.push('event = ?'); params.push(filters.event); }
-  if (filters.webhook_id) { where.push('webhook_id = ?'); params.push(filters.webhook_id); }
-  if (filters.error_only) { where.push('error IS NOT NULL'); }
-  var limit = Math.min(filters.limit || 50, 200);
-  var offset = filters.offset || 0;
-  params.push(limit, offset);
-  return db.prepare('SELECT * FROM webhook_deliveries WHERE ' + where.join(' AND ') + ' ORDER BY created_at DESC LIMIT ? OFFSET ?').all(...params);
-}
-
-export function pruneWebhookDeliveries(keepDays) {
-  var days = keepDays || 7;
-  var result = db.prepare("DELETE FROM webhook_deliveries WHERE created_at < datetime('now', '-' || ? || ' days')").run(days);
-  return result.changes;
-}
-
-// -- Channels --
-
-export function createChannel(name, slug, type, linkedType, linkedId, description, createdBy) {
-  var result = db.prepare(
-    "INSERT INTO channels (name, slug, type, linked_type, linked_id, description, created_by) VALUES (?, ?, ?, ?, ?, ?, ?) RETURNING id"
-  ).get(name, slug, type || 'general', linkedType || null, linkedId || null, description || '', createdBy);
-  return result.id;
-}
-
-export function getChannel(id) {
-  return db.prepare("SELECT * FROM channels WHERE id = ?").get(id);
-}
-
-export function getChannelBySlug(slug) {
-  return db.prepare("SELECT * FROM channels WHERE slug = ?").get(slug);
-}
-
-export function getChannelByLink(linkedType, linkedId) {
-  return db.prepare("SELECT * FROM channels WHERE linked_type = ? AND linked_id = ?").get(linkedType, linkedId);
-}
-
-export function listChannels(filters) {
-  var where = ['1=1'];
-  var params = [];
-  if (filters.type) { where.push('type = ?'); params.push(filters.type); }
-  if (filters.status && filters.status !== 'all') { where.push('status = ?'); params.push(filters.status); }
-  else if (!filters.status) { where.push("status = 'active'"); }
-  if (filters.member) {
-    where.push('id IN (SELECT channel_id FROM channel_members WHERE user_id = ?)');
-    params.push(filters.member);
-  }
-  var limit = Math.min(filters.limit || 50, 500);
-  var offset = filters.offset || 0;
-  params.push(limit, offset);
-  return db.prepare('SELECT * FROM channels WHERE ' + where.join(' AND ') + ' ORDER BY created_at ASC LIMIT ? OFFSET ?').all(...params);
-}
-
-export function updateChannel(id, fields) {
-  buildUpdate('channels', id, fields, ['name', 'description', 'status']);
-}
-
-export function deleteChannel(id) {
-  db.prepare("DELETE FROM channels WHERE id = ?").run(id);
-}
-
-// -- Channel Members --
-
-export function addChannelMember(channelId, userId, userType, role) {
-  try {
-    db.prepare(
-      "INSERT INTO channel_members (channel_id, user_id, user_type, role) VALUES (?, ?, ?, ?)"
-    ).run(channelId, userId, userType || 'agent', role || 'member');
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
-
-export function removeChannelMember(channelId, userId) {
-  var result = db.prepare("DELETE FROM channel_members WHERE channel_id = ? AND user_id = ?").run(channelId, userId);
-  return result.changes > 0;
-}
-
-export function listChannelMembers(channelId) {
-  return db.prepare("SELECT * FROM channel_members WHERE channel_id = ? ORDER BY joined_at ASC").all(channelId);
-}
-
-export function isChannelMember(channelId, userId) {
-  var row = db.prepare("SELECT 1 FROM channel_members WHERE channel_id = ? AND user_id = ?").get(channelId, userId);
-  return !!row;
-}
-
-export function getChannelsByUser(userId) {
-  return db.prepare(
-    "SELECT c.*, cm.role as member_role FROM channels c JOIN channel_members cm ON c.id = cm.channel_id WHERE cm.user_id = ? AND c.status = 'active' ORDER BY c.created_at ASC"
-  ).all(userId);
-}
-
-// -- Channel Read Tracking --
-
-export function markChannelRead(channelId, userId, messageId) {
-  db.prepare(
-    "INSERT INTO channel_reads (channel_id, user_id, last_read_at, last_read_message_id) VALUES (?, ?, datetime('now'), ?) ON CONFLICT(channel_id, user_id) DO UPDATE SET last_read_at = datetime('now'), last_read_message_id = excluded.last_read_message_id"
-  ).run(channelId, userId, messageId || 0);
-}
-
-export function getUnreadCounts(userId) {
-  return db.prepare(
-    "SELECT c.id as channel_id, c.name, c.slug, COUNT(m.id) as unread FROM channels c JOIN channel_members cm ON c.id = cm.channel_id LEFT JOIN messages m ON m.channel_id = c.id AND m.id > COALESCE((SELECT last_read_message_id FROM channel_reads WHERE channel_id = c.id AND user_id = ?), 0) WHERE cm.user_id = ? AND c.status = 'active' GROUP BY c.id"
-  ).all(userId, userId);
-}
-
-export function getLatestChannelMessageId(channelId) {
-  var row = db.prepare("SELECT MAX(id) as max_id FROM messages WHERE channel_id = ?").get(channelId);
-  return row ? (row.max_id || 0) : 0;
-}
-
-// -- Channel Messages --
-
-export function listChannelMessages(channelId, filters) {
-  var where = ['channel_id = ?'];
-  var params = [channelId];
-  if (filters.before) { where.push('id < ?'); params.push(filters.before); }
-  if (filters.after) { where.push('id > ?'); params.push(filters.after); }
-  var limit = Math.min(filters.limit || 50, 500);
-  params.push(limit);
-  return db.prepare(
-    'SELECT * FROM messages WHERE ' + where.join(' AND ') + ' ORDER BY created_at DESC LIMIT ?'
-  ).all(...params);
-}
-
-export function createChannelMessage(channelId, fromAgent, content, metadata) {
-  var result = db.prepare(
-    "INSERT INTO messages (channel_id, from_agent, content, metadata, msg_type) VALUES (?, ?, ?, ?, 'message') RETURNING id"
-  ).get(channelId, fromAgent, content, metadata || '{}');
-  return result.id;
-}
-
-// -- Channel Seeding + Auto-Creation --
-
-export function ensureDefaultChannels() {
-  var defaults = [
-    { name: '#general', slug: 'general', type: 'general', description: 'General discussion' },
-    { name: '#admin', slug: 'admin', type: 'announcement', description: 'Admin coordination' }
-  ];
-  for (var def of defaults) {
-    var existing = getChannelBySlug(def.slug);
-    if (!existing) {
-      var id = createChannel(def.name, def.slug, def.type, null, null, def.description, 'system');
-      var operators = listOperators();
-      for (var op of operators) {
-        addChannelMember(id, op.id, 'operator', 'admin');
-      }
-      if (def.slug === 'general') {
-        var agents = listAgents();
-        for (var agent of agents) {
-          addChannelMember(id, agent.id, 'agent', 'member');
-        }
-      }
-    }
-  }
-}
-
-export function autoCreateEntityChannel(linkedType, linkedId, name, createdBy, memberIds) {
-  var slug = linkedType + '-' + linkedId;
-  var existing = getChannelBySlug(slug);
-  if (existing) return existing.id;
-  var channelName = name || '#' + slug;
-  var id = createChannel(channelName, slug, linkedType, linkedType, linkedId, '', createdBy);
-  var operators = listOperators();
-  for (var op of operators) {
-    addChannelMember(id, op.id, 'operator', 'admin');
-  }
-  if (memberIds) {
-    for (var mid of memberIds) {
-      if (mid) addChannelMember(id, mid, 'agent', 'member');
-    }
-  }
-  return id;
-}
-
-export function getOrCreateDmChannel(userA, userB, userAType, userBType) {
-  // Find any existing active DM channel that has exactly these two participants (case-insensitive).
-  // This prevents duplicate channels when usernames differ by case or creation order.
-  var existing = db.prepare(`
-    SELECT c.id FROM channels c
-    WHERE c.type = 'dm' AND c.status = 'active'
-      AND (SELECT COUNT(*) FROM channel_members m
-           WHERE m.channel_id = c.id AND LOWER(m.user_id) IN (LOWER(?), LOWER(?))) = 2
-      AND (SELECT COUNT(*) FROM channel_members m WHERE m.channel_id = c.id) = 2
-    ORDER BY c.id ASC
-    LIMIT 1
-  `).get(userA, userB);
-  if (existing) {
-    // Ensure both users are members (handles legacy channels missing a member row)
-    addChannelMember(existing.id, userA, userAType || 'agent', 'member');
-    addChannelMember(existing.id, userB, userBType || 'agent', 'member');
-    return existing.id;
-  }
-  // Create new DM channel — use case-insensitive sort for canonical slug
-  var sorted = [userA, userB].sort(function (a, b) {
-    return a.toLowerCase() < b.toLowerCase() ? -1 : a.toLowerCase() > b.toLowerCase() ? 1 : 0;
-  });
-  var slug = 'dm-' + sorted[0].toLowerCase() + '-' + sorted[1].toLowerCase();
-  // Check slug in case a channel exists that our member query didn't catch (e.g. single-member DM)
-  var bySlug = getChannelBySlug(slug);
-  if (bySlug) {
-    addChannelMember(bySlug.id, userA, userAType || 'agent', 'member');
-    addChannelMember(bySlug.id, userB, userBType || 'agent', 'member');
-    return bySlug.id;
-  }
-  var id = createChannel('DM: ' + sorted[0] + ' & ' + sorted[1], slug, 'dm', null, null, '', userA);
-  addChannelMember(id, userA, userAType || 'agent', 'member');
-  addChannelMember(id, userB, userBType || 'agent', 'member');
-  return id;
-}
-
 // -- Drone Jobs --
 
 export function createDroneJob(title, command, inputData, requires, requester, priority, workspaceRepo, workspaceBranch, profileId) {
@@ -2223,532 +1817,6 @@ export function getSlimOverview() {
   return { agents: agents, counts: counts, attention: attention, recent_activity: recent_activity };
 }
 
-// =============== PLUGINS ===============
-
-export function ensurePluginRecord(manifest) {
-  var existing = stmt('dvGetPlugin', 'SELECT * FROM plugins WHERE name = ?').get(manifest.name);
-  if (existing) {
-    stmt('dvUpdatePlugin', `UPDATE plugins SET display_name = ?, description = ?, version = ?, author = ?, route_prefix = ?, mcp_tool_count = ?, updated_at = datetime('now')
-      WHERE name = ?`).run(manifest.displayName || '', manifest.description || '', manifest.version || '1.0.0', manifest.author || '', manifest.routePrefix || '', manifest.mcpToolCount || 0, manifest.name);
-    return { ...existing, updated: true };
-  }
-  // First insert honors the manifest's declared initial state (a plugin that
-  // ships enabled:true loads on first boot); thereafter the existing-record
-  // path above preserves the operator's enable/disable override. Manifests
-  // without an explicit enabled flag still default to disabled (safe).
-  stmt('dvInsertPlugin', `INSERT INTO plugins (name, display_name, description, version, author, enabled, route_prefix, mcp_tool_count)
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?)`).run(manifest.name, manifest.displayName || '', manifest.description || '', manifest.version || '1.0.0', manifest.author || '', manifest.enabled === true ? 1 : 0, manifest.routePrefix || '', manifest.mcpToolCount || 0);
-  return { name: manifest.name, created: true };
-}
-
-export function getPluginRecord(name) {
-  return stmt('dvGetPlugin', 'SELECT * FROM plugins WHERE name = ?').get(name);
-}
-
-export function listPluginRecords() {
-  return db.prepare('SELECT * FROM plugins ORDER BY name').all();
-}
-
-export function updatePluginEnabled(name, enabled) {
-  return db.prepare("UPDATE plugins SET enabled = ?, updated_at = datetime('now') WHERE name = ?").run(enabled ? 1 : 0, name);
-}
-
-export function getPluginMigrationVersion(pluginName) {
-  var row = db.prepare('SELECT MAX(version) as v FROM plugin_migrations WHERE plugin_name = ?').get(pluginName);
-  return row ? (row.v || 0) : 0;
-}
-
-export function recordPluginMigration(pluginName, version, description) {
-  db.prepare('INSERT INTO plugin_migrations (plugin_name, version, description) VALUES (?, ?, ?)').run(pluginName, version, description || '');
-}
-
-// ======== PLUGIN CONFIG ========
-
-export function getPluginConfig(pluginName) {
-  var rows = db.prepare('SELECT key, value, is_secret FROM plugin_config WHERE plugin_name = ?').all(pluginName);
-  return rows;
-}
-
-export function getPluginConfigValue(pluginName, key) {
-  var row = db.prepare('SELECT value FROM plugin_config WHERE plugin_name = ? AND key = ?').get(pluginName, key);
-  return row ? row.value : null;
-}
-
-export function setPluginConfig(pluginName, key, value, isSecret) {
-  db.prepare(
-    `INSERT INTO plugin_config (plugin_name, key, value, is_secret, updated_at) VALUES (?, ?, ?, ?, datetime('now'))
-     ON CONFLICT(plugin_name, key) DO UPDATE SET value = excluded.value, is_secret = excluded.is_secret, updated_at = excluded.updated_at`
-  ).run(pluginName, key, String(value), isSecret ? 1 : 0);
-}
-
-export function deletePluginConfig(pluginName, key) {
-  db.prepare('DELETE FROM plugin_config WHERE plugin_name = ? AND key = ?').run(pluginName, key);
-}
-
-// ======== AGENT SAVEPOINTS ========
-
-export function createSavepoint(agentId, data) {
-  // Validate state_snapshot is valid JSON
-  var stateSnapshot = '{}';
-  if (data.state_snapshot) {
-    try {
-      stateSnapshot = typeof data.state_snapshot === 'string'
-        ? (JSON.parse(data.state_snapshot), data.state_snapshot)
-        : JSON.stringify(data.state_snapshot);
-    } catch (e) {
-      console.warn('[mycelium] Invalid state_snapshot JSON for %s, using empty object', agentId);
-      stateSnapshot = '{}';
-    }
-  }
-
-  var messagesAcked = '[]';
-  if (data.messages_acked) {
-    try {
-      messagesAcked = typeof data.messages_acked === 'string'
-        ? (JSON.parse(data.messages_acked), data.messages_acked)
-        : JSON.stringify(data.messages_acked);
-    } catch (e) {
-      messagesAcked = '[]';
-    }
-  }
-
-  return db.prepare(
-    `INSERT INTO agent_savepoints (agent_id, session_id, heartbeat_at, working_on, state_snapshot, messages_acked, context_versions, notes)
-     VALUES (?, ?, datetime('now'), ?, ?, ?, ?, ?)`
-  ).run(
-    agentId,
-    data.session_id || null,
-    data.working_on || '',
-    stateSnapshot,
-    messagesAcked,
-    JSON.stringify(data.context_versions || {}),
-    data.notes || null
-  );
-}
-
-export function getLatestSavepoint(agentId) {
-  return db.prepare(
-    'SELECT * FROM agent_savepoints WHERE agent_id = ? ORDER BY heartbeat_at DESC LIMIT 1'
-  ).get(agentId);
-}
-
-export function getSavepointHistory(agentId, limit) {
-  return db.prepare(
-    'SELECT id, agent_id, session_id, heartbeat_at, working_on, notes, created_at FROM agent_savepoints WHERE agent_id = ? ORDER BY heartbeat_at DESC LIMIT ?'
-  ).all(agentId, limit || 10);
-}
-
-export function updateSavepointNotes(agentId, notes) {
-  var latest = getLatestSavepoint(agentId);
-  if (!latest) return null;
-  db.prepare('UPDATE agent_savepoints SET notes = ? WHERE id = ?').run(notes, latest.id);
-  return latest.id;
-}
-
-export function computeSavepointDiff(agentId) {
-  var savepoint = getLatestSavepoint(agentId);
-  if (!savepoint) return { has_savepoint: false };
-
-  var ackedIds = [];
-  try { ackedIds = JSON.parse(savepoint.messages_acked || '[]'); } catch (e) { console.warn('[mycelium] JSON parse failed for savepoint.messages_acked (agent: ' + agentId + '):', e.message); }
-
-  var ctxVersions = {};
-  try { ctxVersions = JSON.parse(savepoint.context_versions || '{}'); } catch (e) { console.warn('[mycelium] JSON parse failed for savepoint.context_versions (agent: ' + agentId + '):', e.message); }
-
-  var snapshot = {};
-  try { snapshot = JSON.parse(savepoint.state_snapshot || '{}'); } catch (e) { console.warn('[mycelium] JSON parse failed for savepoint.state_snapshot (agent: ' + agentId + '):', e.message); }
-
-  // Messages the agent hasn't seen (not in acked list, sent after savepoint)
-  var newMessages = db.prepare(
-    "SELECT * FROM messages WHERE (to_agent = ? OR to_agent IS NULL) AND id NOT IN (SELECT value FROM json_each(?)) AND created_at > ? ORDER BY created_at ASC LIMIT 100"
-  ).all(agentId, JSON.stringify(ackedIds), savepoint.heartbeat_at);
-
-  // Tasks that changed since savepoint
-  var tasksChanged = db.prepare(
-    "SELECT * FROM tasks WHERE (assignee = ? OR assignee IS NULL) AND updated_at > ? ORDER BY updated_at DESC LIMIT 50"
-  ).all(agentId, savepoint.heartbeat_at);
-
-  // Context keys that changed since savepoint
-  var contextChanged = db.prepare(
-    "SELECT * FROM context_keys WHERE updated_at > ? ORDER BY updated_at DESC LIMIT 50"
-  ).all(savepoint.heartbeat_at);
-
-  // Plans that changed since savepoint
-  var plansChanged = db.prepare(
-    "SELECT p.* FROM plans p WHERE p.updated_at > ? ORDER BY p.updated_at DESC LIMIT 20"
-  ).all(savepoint.heartbeat_at);
-
-  // Bugs that changed since savepoint
-  var bugsChanged = db.prepare(
-    "SELECT * FROM bugs WHERE updated_at > ? ORDER BY updated_at DESC LIMIT 20"
-  ).all(savepoint.heartbeat_at);
-
-  // Drone jobs that changed since savepoint
-  var droneJobsChanged = db.prepare(
-    "SELECT * FROM drone_jobs WHERE (started_at > ? OR completed_at > ? OR created_at > ?) ORDER BY created_at DESC LIMIT 20"
-  ).all(savepoint.heartbeat_at, savepoint.heartbeat_at, savepoint.heartbeat_at);
-
-  // Events since savepoint
-  var eventsSince = db.prepare(
-    "SELECT * FROM events WHERE created_at > ? ORDER BY created_at DESC LIMIT 50"
-  ).all(savepoint.heartbeat_at);
-
-  return {
-    has_savepoint: true,
-    savepoint_id: savepoint.id,
-    savepoint_at: savepoint.heartbeat_at,
-    session_id: savepoint.session_id,
-    was_working_on: savepoint.working_on,
-    notes: savepoint.notes,
-    previous_state: snapshot,
-    changes: {
-      new_messages: newMessages,
-      tasks_changed: tasksChanged,
-      context_changed: contextChanged,
-      plans_changed: plansChanged,
-      bugs_changed: bugsChanged,
-      drone_jobs_changed: droneJobsChanged,
-      events_since: eventsSince.length
-    },
-    summary: {
-      messages: newMessages.length,
-      tasks: tasksChanged.length,
-      context: contextChanged.length,
-      plans: plansChanged.length,
-      bugs: bugsChanged.length,
-      drone_jobs: droneJobsChanged.length,
-      events: eventsSince.length,
-      time_since: savepoint.heartbeat_at
-    }
-  };
-}
-
-export function pruneSavepoints(agentId, keepCount) {
-  // Keep only the most recent N savepoints per agent
-  var count = keepCount || 50;
-  var cutoff = db.prepare(
-    'SELECT heartbeat_at FROM agent_savepoints WHERE agent_id = ? ORDER BY heartbeat_at DESC LIMIT 1 OFFSET ?'
-  ).get(agentId, count);
-  if (cutoff) {
-    db.prepare('DELETE FROM agent_savepoints WHERE agent_id = ? AND heartbeat_at < ?').run(agentId, cutoff.heartbeat_at);
-  }
-}
-
-// =============== NODE PROFILES — Stand Up Calibration ===============
-
-var NODE_PROFILE_JSON_FIELDS = ['rules', 'required_concepts', 'mcp_config', 'tool_whitelist', 'repo_list', 'md_checkpoints', 'md_blocklist'];
-
-function stringifyProfileField(val) {
-  if (val === undefined || val === null) return undefined;
-  if (typeof val === 'string') return val;
-  return JSON.stringify(val);
-}
-
-function parseProfileRow(row) {
-  if (!row) return null;
-  for (var f of NODE_PROFILE_JSON_FIELDS) {
-    if (row[f]) {
-      try { row[f] = JSON.parse(row[f]); } catch (e) { /* keep as string */ }
-    }
-  }
-  return row;
-}
-
-export function createNodeProfile(id, data) {
-  var d = data || {};
-  db.prepare(
-    'INSERT INTO node_profiles (id, node_type, layer, parent_id, rules, required_concepts, mcp_config, tool_whitelist, repo_list, md_checkpoints, md_blocklist) ' +
-    'VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)'
-  ).run(
-    id,
-    d.node_type || 'agent',
-    d.layer || 'customer',
-    d.parent_id || null,
-    stringifyProfileField(d.rules) || '{}',
-    stringifyProfileField(d.required_concepts) || '[]',
-    stringifyProfileField(d.mcp_config) || '{}',
-    stringifyProfileField(d.tool_whitelist) || '[]',
-    stringifyProfileField(d.repo_list) || '[]',
-    stringifyProfileField(d.md_checkpoints) || '[]',
-    stringifyProfileField(d.md_blocklist) || '[]'
-  );
-  return getNodeProfile(id);
-}
-
-export function getNodeProfile(id) {
-  var row = db.prepare('SELECT * FROM node_profiles WHERE id = ?').get(id);
-  return parseProfileRow(row);
-}
-
-export function listNodeProfiles(filter) {
-  var where = [];
-  var params = [];
-  if (filter && filter.node_type) { where.push('node_type = ?'); params.push(filter.node_type); }
-  if (filter && filter.layer) { where.push('layer = ?'); params.push(filter.layer); }
-  var sql = 'SELECT * FROM node_profiles' + (where.length ? ' WHERE ' + where.join(' AND ') : '') + ' ORDER BY layer, node_type, id';
-  var rows = db.prepare(sql).all.apply(db.prepare(sql), params);
-  return rows.map(parseProfileRow);
-}
-
-export function updateNodeProfile(id, data) {
-  var existing = getNodeProfile(id);
-  if (!existing) return null;
-  if (existing.layer === 'platform') return null;
-
-  var sets = ["updated_at = datetime('now')"];
-  var values = [];
-  if (data.node_type !== undefined) { sets.push('node_type = ?'); values.push(data.node_type); }
-  if (data.layer !== undefined) { sets.push('layer = ?'); values.push(data.layer); }
-  if (data.parent_id !== undefined) { sets.push('parent_id = ?'); values.push(data.parent_id); }
-  for (var f of NODE_PROFILE_JSON_FIELDS) {
-    if (data[f] !== undefined) { sets.push(f + ' = ?'); values.push(stringifyProfileField(data[f])); }
-  }
-  if (values.length === 0) return existing;
-  values.push(id);
-  db.prepare('UPDATE node_profiles SET ' + sets.join(', ') + ' WHERE id = ?').run.apply(
-    db.prepare('UPDATE node_profiles SET ' + sets.join(', ') + ' WHERE id = ?'), values
-  );
-  return getNodeProfile(id);
-}
-
-export function deleteNodeProfile(id) {
-  var existing = getNodeProfile(id);
-  if (!existing) return null;
-  if (existing.layer === 'platform') return null;
-  db.prepare('DELETE FROM node_profiles WHERE id = ?').run(id);
-  return existing;
-}
-
-export function resolveProfileChain(agentId) {
-  // Load agent to determine type
-  var agent = getAgent(agentId);
-  var agentType = (agent && agent.agent_type) ? agent.agent_type : 'agent';
-
-  // Build chain: platform default -> customer default -> agent-specific
-  var chainIds = [
-    'default-' + agentType,
-    'customer-' + agentType,
-    agentId
-  ];
-
-  var merged = {
-    rules: {},
-    required_concepts: [],
-    mcp_config: {},
-    tool_whitelist: [],
-    repo_list: [],
-    md_checkpoints: [],
-    md_blocklist: [],
-    layers_applied: []
-  };
-
-  // Track platform critical rules so they can't be downgraded
-  var platformCritical = {};
-
-  for (var profileId of chainIds) {
-    var profile = getNodeProfile(profileId);
-    if (!profile) continue;
-
-    merged.layers_applied.push({ id: profile.id, layer: profile.layer, node_type: profile.node_type });
-
-    // Merge rules: later layers override, but can't downgrade platform critical severity
-    var rules = profile.rules || {};
-    for (var ruleKey in rules) {
-      if (profile.layer === 'platform' && rules[ruleKey].severity === 'critical') {
-        platformCritical[ruleKey] = true;
-      }
-      if (platformCritical[ruleKey] && profile.layer !== 'platform') {
-        // Can't downgrade platform critical — keep severity, allow other fields to merge
-        var incoming = typeof rules[ruleKey] === 'object' ? Object.assign({}, rules[ruleKey]) : { severity: rules[ruleKey] };
-        incoming.severity = 'critical';
-        merged.rules[ruleKey] = incoming;
-      } else {
-        merged.rules[ruleKey] = rules[ruleKey];
-      }
-    }
-
-    // Overlay objects: mcp_config
-    var mcp = profile.mcp_config || {};
-    for (var mk in mcp) {
-      merged.mcp_config[mk] = mcp[mk];
-    }
-
-    // Concatenate + deduplicate arrays
-    var arrayFields = ['required_concepts', 'tool_whitelist', 'repo_list', 'md_checkpoints', 'md_blocklist'];
-    for (var af of arrayFields) {
-      var arr = profile[af];
-      if (Array.isArray(arr)) {
-        for (var item of arr) {
-          if (merged[af].indexOf(item) === -1) {
-            merged[af].push(item);
-          }
-        }
-      }
-    }
-  }
-
-  return merged;
-}
-
-// ---- Seed Platform Profiles ----
-
-export function seedPlatformProfiles() {
-  // Only seed if default-agent doesn't exist yet
-  var existing = db.prepare('SELECT id FROM node_profiles WHERE id = ?').get('default-agent');
-  if (existing) return;
-
-  // default-agent: base rules for all agents
-  db.prepare(
-    'INSERT INTO node_profiles (id, node_type, layer, parent_id, rules, required_concepts, mcp_config, tool_whitelist, repo_list, md_checkpoints, md_blocklist) ' +
-    'VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)'
-  ).run(
-    'default-agent',
-    'agent',
-    'platform',
-    null,
-    JSON.stringify({
-      honesty: { severity: 'critical', description: 'Never fabricate information or pretend something worked when it did not' },
-      evidence_based: { severity: 'critical', description: 'Verify files exist before editing. Read before writing. No guessing.' },
-      identity: { severity: 'high', description: 'Maintain assigned agent identity. Do not impersonate other agents.' },
-      communication: { severity: 'high', description: 'Report failures immediately. Use clear, direct language.' },
-      coordination: { severity: 'high', description: 'Update the network when fixing bugs, completing steps, or assigning work.' },
-      security: { severity: 'critical', description: 'Never commit secrets, credentials, or API keys. Never expose admin keys.' },
-      paid_services: { severity: 'critical', description: 'Never call paid APIs without explicit approval. No unauthorized spending.' },
-      code_standards: { severity: 'medium', description: 'Follow existing codebase conventions. Match style of surrounding code.' }
-    }),
-    '[]',
-    '{}',
-    '[]',
-    '[]',
-    JSON.stringify(['mycelium_boot', 'No guessing', 'No silent failures']),
-    JSON.stringify(['studio_boot', 'studio_get_work', 'studio_read_messages', 'generate_sprites.py', 'Pixel Arena', 'Some of You May Die'])
-  );
-
-  // default-drone: minimal rules for GPU/CPU workers
-  db.prepare(
-    'INSERT INTO node_profiles (id, node_type, layer, parent_id, rules, required_concepts, mcp_config, tool_whitelist, repo_list, md_checkpoints, md_blocklist) ' +
-    'VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)'
-  ).run(
-    'default-drone',
-    'drone',
-    'platform',
-    null,
-    JSON.stringify({
-      execution: { severity: 'critical', description: 'Execute assigned jobs faithfully. Report results accurately.' },
-      no_messages: { severity: 'high', description: 'Drones do not send messages or participate in coordination. Execute only.' }
-    }),
-    '[]',
-    '{}',
-    '[]',
-    '[]',
-    '[]',
-    '[]'
-  );
-
-  // default-admin: inherits agent rules + coordination emphasis
-  db.prepare(
-    'INSERT INTO node_profiles (id, node_type, layer, parent_id, rules, required_concepts, mcp_config, tool_whitelist, repo_list, md_checkpoints, md_blocklist) ' +
-    'VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)'
-  ).run(
-    'default-admin',
-    'admin',
-    'platform',
-    'default-agent',
-    JSON.stringify({
-      honesty: { severity: 'critical', description: 'Never fabricate information or pretend something worked when it did not' },
-      evidence_based: { severity: 'critical', description: 'Verify files exist before editing. Read before writing. No guessing.' },
-      identity: { severity: 'high', description: 'Maintain assigned agent identity. Do not impersonate other agents.' },
-      communication: { severity: 'high', description: 'Report failures immediately. Use clear, direct language.' },
-      coordination: { severity: 'critical', description: 'Coordinate work across all agents. Ensure no agent is idle. Route tasks by domain.' },
-      security: { severity: 'critical', description: 'Never commit secrets, credentials, or API keys. Never expose admin keys.' },
-      paid_services: { severity: 'critical', description: 'Never call paid APIs without explicit approval. No unauthorized spending.' },
-      code_standards: { severity: 'medium', description: 'Follow existing codebase conventions. Match style of surrounding code.' }
-    }),
-    '[]',
-    '{}',
-    '[]',
-    '[]',
-    JSON.stringify(['mycelium_boot', 'No guessing', 'No silent failures']),
-    JSON.stringify(['studio_boot', 'studio_get_work', 'studio_read_messages', 'generate_sprites.py', 'Pixel Arena', 'Some of You May Die'])
-  );
-
-  console.log('Seeded platform node profiles: default-agent, default-drone, default-admin');
-}
-
-// ---- Stand Up: Calibration Block ----
-
-export function buildCalibrationBlock(agentId) {
-  var resolved = resolveProfileChain(agentId);
-  var drift = [];
-
-  // Get latest savepoint to extract md_report
-  var savepoint = getLatestSavepoint(agentId);
-  var stateSnapshot = {};
-  if (savepoint && savepoint.state_snapshot) {
-    if (typeof savepoint.state_snapshot === 'object') {
-      stateSnapshot = savepoint.state_snapshot;
-    } else {
-      try { stateSnapshot = JSON.parse(savepoint.state_snapshot || '{}'); } catch (e) { /* */ }
-    }
-  }
-
-  var mdReport = stateSnapshot.md_report || null;
-
-  // Also check context key for md_report (heartbeat may have persisted it)
-  if (!mdReport) {
-    var mdCtx = getContextKey(agentId, 'md_report');
-    if (mdCtx && mdCtx.data) {
-      try { mdReport = typeof mdCtx.data === 'object' ? mdCtx.data : JSON.parse(mdCtx.data); } catch (e) { /* */ }
-    }
-  }
-
-  if (mdReport) {
-    // Check anchors: md_checkpoints should be present in agent's CLAUDE.md
-    var checkpoints = resolved.md_checkpoints || [];
-    var anchorsPresent = mdReport.anchors_present || [];
-    for (var i = 0; i < checkpoints.length; i++) {
-      var cp = checkpoints[i];
-      if (anchorsPresent.indexOf(cp) === -1) {
-        drift.push({ level: 'warning', rule: 'md_checkpoint_missing', detail: 'Expected anchor not found in CLAUDE.md: ' + cp });
-      }
-    }
-
-    // Check blocklist: md_blocklist items should NOT be present
-    var blocklist = resolved.md_blocklist || [];
-    var blocklistFound = mdReport.blocklist_found || [];
-    for (var j = 0; j < blocklist.length; j++) {
-      var bl = blocklist[j];
-      if (blocklistFound.indexOf(bl) !== -1) {
-        drift.push({ level: 'critical', rule: 'md_blocklist_found', detail: 'Blocked term found in CLAUDE.md: ' + bl });
-      }
-    }
-  } else {
-    drift.push({ level: 'info', rule: 'md_report_missing', detail: 'Send md_report in heartbeat state_snapshot to enable CLAUDE.md drift detection' });
-  }
-
-  // Determine status based on drift items
-  var status = 'aligned';
-  for (var d = 0; d < drift.length; d++) {
-    if (drift[d].level === 'critical') { status = 'critical'; break; }
-    if (drift[d].level === 'warning') { status = 'drifted'; }
-  }
-
-  var calibration = {
-    status: status,
-    profile_chain: resolved.layers_applied || [],
-    rules: resolved.rules || {},
-    drift: drift,
-    md_checkpoints: resolved.md_checkpoints || [],
-    md_blocklist: resolved.md_blocklist || [],
-    last_standup: new Date().toISOString()
-  };
-
-  // Persist to context key
-  upsertContextKey(agentId, 'standup', JSON.stringify(calibration), 'system');
-
-  return calibration;
-}
-
 // =============== TEAM SETTINGS ===============
 
 export function listTeamSettings(section) {
@@ -2972,188 +2040,3 @@ export function getTeamProjects(teamId) {
 }
 
 
-// -- Agent Profiles --
-
-export function getAgentProfile(agentId) {
-  var row = db.prepare('SELECT * FROM agent_profiles WHERE agent_id = ?').get(agentId);
-  if (row) {
-    try { row.specializations = JSON.parse(row.specializations); } catch (e) { row.specializations = []; }
-    try { row.preferred_projects = JSON.parse(row.preferred_projects); } catch (e) { row.preferred_projects = []; }
-    try { row.capability_history = JSON.parse(row.capability_history); } catch (e) { row.capability_history = []; }
-    try { row.profile_data = JSON.parse(row.profile_data); } catch (e) { row.profile_data = {}; }
-  }
-  return row;
-}
-
-export function ensureAgentProfile(agentId) {
-  var existing = db.prepare('SELECT agent_id FROM agent_profiles WHERE agent_id = ?').get(agentId);
-  if (existing) {
-    db.prepare("UPDATE agent_profiles SET session_count = session_count + 1, last_active_at = datetime('now') WHERE agent_id = ?").run(agentId);
-    return getAgentProfile(agentId);
-  }
-  var agent = getAgent(agentId);
-  var displayName = agent ? agent.name : agentId;
-  db.prepare("INSERT INTO agent_profiles (agent_id, display_name, session_count) VALUES (?, ?, 1)").run(agentId, displayName);
-  return getAgentProfile(agentId);
-}
-
-export function updateAgentProfile(agentId, fields) {
-  var f = Object.assign({}, fields);
-  if (f.specializations !== undefined) f.specializations = JSON.stringify(f.specializations);
-  if (f.preferred_projects !== undefined) f.preferred_projects = JSON.stringify(f.preferred_projects);
-  if (f.profile_data !== undefined) f.profile_data = JSON.stringify(f.profile_data);
-  if (f.capability_history !== undefined) f.capability_history = JSON.stringify(f.capability_history);
-  var changed = buildUpdate('agent_profiles', agentId, f, ['display_name', 'specializations', 'preferred_projects', 'max_concurrent', 'profile_data', 'capability_history'], { extraSets: ["last_active_at = datetime('now')"], where: 'agent_id' });
-  if (!changed) return getAgentProfile(agentId);
-  return getAgentProfile(agentId);
-}
-
-export function incrementProfileCounter(agentId, counter) {
-  var allowed = ['total_tasks_completed', 'total_bugs_fixed', 'total_prs_created'];
-  if (allowed.indexOf(counter) === -1) return;
-  db.prepare('UPDATE agent_profiles SET ' + counter + ' = ' + counter + ' + 1 WHERE agent_id = ?').run(agentId);
-}
-
-export function listAgentProfiles() {
-  var rows = db.prepare('SELECT * FROM agent_profiles ORDER BY total_tasks_completed DESC').all();
-  return rows.map(function (row) {
-    try { row.specializations = JSON.parse(row.specializations); } catch (e) { row.specializations = []; }
-    try { row.preferred_projects = JSON.parse(row.preferred_projects); } catch (e) { row.preferred_projects = []; }
-    try { row.capability_history = JSON.parse(row.capability_history); } catch (e) { row.capability_history = []; }
-    try { row.profile_data = JSON.parse(row.profile_data); } catch (e) { row.profile_data = {}; }
-    return row;
-  });
-}
-
-export function getAgentLeaderboard(limit) {
-  limit = Math.min(limit || 20, 100);
-  var rows = db.prepare(
-    'SELECT agent_id, display_name, specializations, total_tasks_completed, total_bugs_fixed, total_prs_created, session_count, first_seen_at, last_active_at FROM agent_profiles ORDER BY total_tasks_completed DESC LIMIT ?'
-  ).all(limit);
-  return rows.map(function (row) {
-    try { row.specializations = JSON.parse(row.specializations); } catch (e) { row.specializations = []; }
-    return row;
-  });
-}
-
-// -- Health Patrol --
-
-// Reconciliation read-surface (A7 — state-desync visibility).
-//
-// Health Patrol's getStaleTasks/getStalePlanSteps key staleness off the
-// ASSIGNEE'S heartbeat — they answer "is the worker gone?". Reconciliation
-// answers a different question: "has this RECORD itself drifted from reality?"
-// i.e. it has sat in_progress past a threshold with no edits, regardless of
-// whether any agent is online. That's the silently-stuck case — e.g. all 12
-// bugs sitting in_progress including #10 (fixed-in-code) — that corrupts
-// greenlight decisions and training labels.
-//
-// READ-ONLY. Returns candidate records; never mutates. The operator decides
-// what to reconcile. thresholdMinutes defaults to 24h (1440).
-export function getReconciliationCandidates(thresholdMinutes) {
-  thresholdMinutes = thresholdMinutes || (24 * 60);
-  var cutoffClause = "updated_at < datetime('now', '-' || ? || ' minutes')";
-
-  var bugs = db.prepare(
-    "SELECT id, project_id, title, status, assignee, updated_at FROM bugs " +
-    "WHERE status = 'in_progress' AND " + cutoffClause + " ORDER BY updated_at ASC"
-  ).all(thresholdMinutes);
-
-  var tasks = db.prepare(
-    "SELECT id, project_id, title, status, assignee, updated_at FROM tasks " +
-    "WHERE status = 'in_progress' AND " + cutoffClause + " ORDER BY updated_at ASC"
-  ).all(thresholdMinutes);
-
-  var planSteps = db.prepare(
-    "SELECT s.id, s.plan_id, s.title, s.status, s.assignee, s.updated_at FROM plan_steps s " +
-    "WHERE s.status = 'in_progress' AND s." + cutoffClause + " ORDER BY s.updated_at ASC"
-  ).all(thresholdMinutes);
-
-  return {
-    threshold_minutes: thresholdMinutes,
-    generated_at: db.prepare("SELECT datetime('now') as now").get().now,
-    counts: { bugs: bugs.length, tasks: tasks.length, plan_steps: planSteps.length,
-              total: bugs.length + tasks.length + planSteps.length },
-    bugs: bugs,
-    tasks: tasks,
-    plan_steps: planSteps
-  };
-}
-
-export function getStaleAgents(thresholdMinutes) {
-  thresholdMinutes = thresholdMinutes || 15;
-  return db.prepare(
-    "SELECT id, name, status, working_on, last_heartbeat FROM agents WHERE status IN ('online', 'idle', 'busy') AND last_heartbeat < datetime('now', '-' || ? || ' minutes') AND role != 'drone'"
-  ).all(thresholdMinutes);
-}
-
-export function getStaleTasks(thresholdMinutes) {
-  thresholdMinutes = thresholdMinutes || 30;
-  return db.prepare(
-    "SELECT t.id, t.title, t.assignee, t.updated_at FROM tasks t LEFT JOIN agents a ON t.assignee = a.id WHERE t.status = 'in_progress' AND (a.last_heartbeat IS NULL OR a.last_heartbeat < datetime('now', '-' || ? || ' minutes'))"
-  ).all(thresholdMinutes);
-}
-
-export function getStaleRequests(thresholdMinutes) {
-  thresholdMinutes = thresholdMinutes || 60;
-  return db.prepare(
-    "SELECT id, from_agent, to_agent, content, created_at FROM messages WHERE msg_type = 'request' AND status IN ('sent', 'pending') AND created_at < datetime('now', '-' || ? || ' minutes')"
-  ).all(thresholdMinutes);
-}
-
-export function getStaleDrones(thresholdMinutes) {
-  thresholdMinutes = thresholdMinutes || 30;
-  return db.prepare(
-    "SELECT id, name, status, last_heartbeat FROM agents WHERE (role = 'drone' OR project_id = 'drone') AND status IN ('online', 'idle') AND last_heartbeat < datetime('now', '-' || ? || ' minutes')"
-  ).all(thresholdMinutes);
-}
-
-export function getStalePlanSteps(thresholdMinutes) {
-  thresholdMinutes = thresholdMinutes || 120;
-  return db.prepare(
-    "SELECT s.id, s.title, s.assignee, s.plan_id, s.updated_at FROM plan_steps s JOIN plans p ON p.id = s.plan_id WHERE s.status = 'in_progress' AND p.status = 'active' AND s.updated_at < datetime('now', '-' || ? || ' minutes')"
-  ).all(thresholdMinutes);
-}
-
-
-// ---- Agent Templates ----
-
-export function createAgentTemplate(id, name, description, data, createdBy) {
-  db.prepare(
-    "INSERT INTO agent_templates (id, name, description, runtime, llm_backend, llm_model, agent_type, capabilities, project_id, team_ids, profile_rules, created_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"
-  ).run(id, name, description || '', data.runtime || '', data.llm_backend || '', data.llm_model || '', data.agent_type || 'agent', JSON.stringify(data.capabilities || ['code', 'assets']), data.project_id || '', JSON.stringify(data.team_ids || []), JSON.stringify(data.profile_rules || {}), createdBy || '');
-  return getAgentTemplate(id);
-}
-
-export function getAgentTemplate(id) {
-  var row = db.prepare("SELECT * FROM agent_templates WHERE id = ?").get(id);
-  if (!row) return null;
-  try { row.capabilities = JSON.parse(row.capabilities); } catch (_) { row.capabilities = []; }
-  try { row.team_ids = JSON.parse(row.team_ids); } catch (_) { row.team_ids = []; }
-  try { row.profile_rules = JSON.parse(row.profile_rules); } catch (_) { row.profile_rules = {}; }
-  return row;
-}
-
-export function listAgentTemplates() {
-  var rows = db.prepare("SELECT * FROM agent_templates ORDER BY name").all();
-  for (var row of rows) {
-    try { row.capabilities = JSON.parse(row.capabilities); } catch (_) { row.capabilities = []; }
-    try { row.team_ids = JSON.parse(row.team_ids); } catch (_) { row.team_ids = []; }
-    try { row.profile_rules = JSON.parse(row.profile_rules); } catch (_) { row.profile_rules = {}; }
-  }
-  return rows;
-}
-
-export function updateAgentTemplate(id, fields) {
-  var f = Object.assign({}, fields);
-  if (f.capabilities !== undefined) f.capabilities = JSON.stringify(f.capabilities);
-  if (f.team_ids !== undefined) f.team_ids = JSON.stringify(f.team_ids);
-  if (f.profile_rules !== undefined) f.profile_rules = JSON.stringify(f.profile_rules);
-  var changed = buildUpdate('agent_templates', id, f, ['name', 'description', 'runtime', 'llm_backend', 'llm_model', 'agent_type', 'project_id', 'capabilities', 'team_ids', 'profile_rules'], { updatedAt: true });
-  if (!changed) return getAgentTemplate(id);
-  return getAgentTemplate(id);
-}
-
-export function deleteAgentTemplate(id) {
-  db.prepare("DELETE FROM agent_templates WHERE id = ?").run(id);
-}
